requirements.txt
tensorflow-gpu == 1.9.0
numpy == 1.14.5
dependency_injector == 3.13.1